Update all the things! (Including Kotlin 1.3)

+++ feature/teams/src/main/java/com/supercilex/robotscouter/feature/teams/Selection.kt
@@ -26,6 +26,6 @@ internal class TeamDetails(
 
     override fun inSelectionHotspot(e: MotionEvent): Boolean {
         val media = holder.media
-        return e.rawX in media.left..media.right
+        return e.rawX.toInt() in media.left..media.right
     }
 }
